refactor usage of live555helper class to avoid duplication of code
